Add option for unix-socket in config.yml
Add configuration option to allow using unix-socket to connect
with the daemonized telegram-cli (i.e.  using the command `telegram-cli -d -S
/var/run/telegram.sock`).

If null value is set for the socket pack the TCP connection is used.
If a path is set for the socket TCP connection options are ignored.

WHY: With unix-socket running telegram-cli as a daemon is more secure.
With tpc connection the interface to user personal data served by telegram-cli
is accessible system wide.
With unix-socket user can set specific privileges on the socket
and thus restrict the access to that interface.